model/PIFu-multiview/requirements.txt
cycler
decorator
gradio
imageio
kiwisolver
matplotlib
networkx
numpy==1.26.4
opencv-python
pathlib
Pillow
PyOpenGL
pyparsing
python-dateutil
PyWavelets
scikit-image
scipy
Shapely
six
torch
torchvision
tqdm
trimesh
xxhash
rembg